flask/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-GraphQL==2.0.1
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
graphql-server-core==1.2.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
promise==2.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
Rx==1.6.3
six==1.16.0
strawberry-graphql==0.247.0
typing_extensions==4.12.2
Werkzeug==3.0.4
